requirements.txt
mido==1.2.9
Pillow==7.0.0
